change all-reports path
